Rename DropApiProvider types
