--- packages/providers/src/ports/DropApiProvider/Types/index.ts
@@ -1,2 +1,0 @@
-export * from './input';
-export * from './response';
