+++ packages/providers/src/ports/DropApiProvider/index.ts
@@ -1,2 +1,2 @@
-export * from './Types';
+export * from './types';
 export { DropApiProvider } from './DropApiProvider';
