+++ packages/providers/src/ports/DropApiProvider/types/index.ts
@@ -1,0 +1,2 @@
+export { CreateDropInput, UpdateDropInput } from './DropInput';
+export { DropResponse } from './DropResponse';
